backend/requirements.txt
alembic==1.11.2
aniso8601==9.0.1
attrs==23.1.0
blinker==1.6.2
click==8.1.5
colorama==0.4.6
easyocr==1.7.2
exceptiongroup==1.1.3
filelock==3.18.0
Flask==2.2.5
Flask-JWT-Extended==4.5.2
Flask-Migrate==4.0.4
flask-restx==1.1.0
Flask-SQLAlchemy==3.0.5
fsspec==2025.5.1
greenlet==2.0.2
gunicorn==21.2.0
idna==3.10
imageio==2.37.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
jsonschema-specifications==0.1.3
lazy_loader==0.4
Mako==1.2.4
MarkupSafe==2.1.3
mpmath==1.3.0
multidict==6.4.4
networkx==3.4.2
ninja==1.11.1.4
numpy==2.2.6
opencv-python-headless==4.11.0.86
packaging==23.2
pdf2image==1.17.0
pillow==11.2.1
pluggy==1.2.0
propcache==0.3.1
psycopg2-binary==2.9.9
pyclipper==1.3.0.post6
PyJWT==2.8.0
pyrsistent==0.20.0
pytesseract==0.3.13
pytest==7.4.2
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==0.21.1
pytz==2023.3
PyYAML==6.0.2
referencing==0.8.9
scikit-image==0.25.2
scipy==1.15.3
shapely==2.1.1
six==1.17.0
SQLAlchemy==2.0.19
sympy==1.14.0
tifffile==2025.5.26
tomli==2.0.1
torch==2.7.0
torchvision==0.22.0
typing_extensions==4.13.2
Werkzeug==2.2.3
yarl==1.20.0
